# internal/apidocs/gen/swagger.json
{
    "components": {
        "schemas": {
            "auth.ErrorMsg": {
                "properties": {
                    "msg": {
                        "description": "Error message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "server.protocolOpenRequest": {
                "properties": {
                    "secret": {
                        "description": "Contents of the running agent's protocol.secret file",
                        "type": "string"
                    }
                },
                "required": [
                    "secret"
                ],
                "type": "object"
            },
            "server.protocolOpenResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "server.swapAreaRow": {
                "properties": {
                    "free_bytes": {
                        "type": "integer"
                    },
                    "host": {
                        "type": "string"
                    },
                    "scan_timestamp": {
                        "type": "string"
                    },
                    "size_bytes": {
                        "type": "integer"
                    },
                    "swapfile": {
                        "type": "string"
                    },
                    "used_bytes": {
                        "type": "integer"
                    },
                    "utilization_pct": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "server.swapAreasPagination": {
                "properties": {
                    "hasMore": {
                        "type": "boolean"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "offset": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "server.swapAreasResponse": {
                "properties": {
                    "pagination": {
                        "$ref": "#/components/schemas/server.swapAreasPagination"
                    },
                    "swapAreas": {
                        "items": {
                            "$ref": "#/components/schemas/server.swapAreaRow"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "totalCount": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "server.trayClaimRequest": {
                "properties": {
                    "token": {
                        "description": "Single-use tray token from the #tray= URL fragment",
                        "type": "string"
                    }
                },
                "required": [
                    "token"
                ],
                "type": "object"
            },
            "server.trayClaimResponse": {
                "properties": {
                    "api_key": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "The swag-generated half of the Agent API v1 document. Serve-time authority lives in internal/apidocs: the hand-maintained fragment (openapi.json) owns info/tags/components/security and every not-yet-migrated path, and this generated document's paths merge into it.",
        "title": "Agent API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/auth/tray-claim": {
            "post": {
                "description": "Public: the token itself is the credential — minted seconds earlier by the local user's physical tray Open click (or an hwa:// protocol invocation), carried in the URL fragment, single-use, 60-second TTL. This is how a desktop user gets a signed-in UI without ever seeing a login screen. Each claim also REAPS older tray-handoff keys beyond the newest 5 (every Open mints a key and nothing else retires them; unbounded piles make the cold-boot bcrypt scan crawl) — a browser tab holding a reaped key re-signs-in via the tray.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.trayClaimRequest",
                                        "summary": "request",
                                        "description": "Tray claim request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Tray claim request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.trayClaimResponse"
                                }
                            }
                        },
                        "description": "Fresh admin key, named after the local OS account"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.ErrorMsg"
                                }
                            }
                        },
                        "description": "Missing token"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.ErrorMsg"
                                }
                            }
                        },
                        "description": "Unknown, expired, or already-used token"
                    }
                },
                "summary": "Exchange a tray one-time token for an admin API key",
                "tags": [
                    "Local Login"
                ]
            }
        },
...
        "/protocol/open": {
            "post": {
                "description": "Public but secret-gated: when the OS spawns a fresh agent process for an hwa://open invocation (Windows registry handler, Linux .desktop handler), that process forwards the action here and exits. The per-boot secret file (0600, beside the running agent's config) authenticates it — web pages cannot read local files, so possession proves a local same-user process. On success the running agent opens the signed-in UI in the user's browser, exactly like a tray Open click.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.protocolOpenRequest",
                                        "summary": "request",
                                        "description": "Protocol open request"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Protocol open request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.protocolOpenResponse"
                                }
                            }
                        },
                        "description": "Action accepted; the agent is opening the browser"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.ErrorMsg"
                                }
                            }
                        },
                        "description": "Missing secret"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.ErrorMsg"
                                }
                            }
                        },
                        "description": "Invalid secret"
                    }
                },
                "summary": "hwa:// single-instance handoff",
                "tags": [
                    "Local Login"
                ]
            }
        },
...
        "/system/swap/areas": {
            "get": {
                "description": "Minimum role: viewer. Row-per-area listing with the Node payload's pagination envelope, read live from the OS. The zvol pool filter has no meaning on this agent and is ignored.",
                "parameters": [
                    {
                        "description": "Page size",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 100,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Page offset",
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "default": 0,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.swapAreasResponse"
                                }
                            }
                        },
                        "description": "Swap areas"
                    }
                },
                "summary": "List swap areas",
                "tags": [
                    "Swap Management"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}